+++ cmd/mindwalk/main.go
@@ -18,6 +18,7 @@ import (
 	"github.com/cosmtrek/mindwalk/internal/model"
 	"github.com/cosmtrek/mindwalk/internal/server"
 )
+
 func main() {
 	if err := run(os.Args[1:]); err != nil {
 		fmt.Fprintln(os.Stderr, "mindwalk:", err)